에디터 붙이기 완료

+++ pages/api/clear.ts
@@ -23,19 +23,22 @@ handler.get(async (req, res) => {
     const serial = req.query.serial
     const path = 'temp/' + serial
     if (serial) {
-      const files = fs.readdirSync(path)
-      await Promise.all(
-        files.map(name => {
-          fs.unlinkSync(path + '/' + name)
-        })
-      )
-      fs.rmdirSync(path)
+      if (fs.existsSync(path) === true) {
+        const files = fs.readdirSync(path)
+        if (files && files.length > 0) {
+          await Promise.all(
+            files.map(name => {
+              fs.unlinkSync(path + '/' + name)
+            })
+          )
+        }
+        fs.rmdirSync(path)
+      }
       res.status(200).json({ result: true })
     } else {
       res.status(400).json({ result: false, message: 'Sorry!' })
     }
   } catch (e) {
-    console.log(e)
     res.status(400).json({ result: false, message: 'Sorry!' })
   }
 })
